% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoiextent.R
\name{aoiextent}
\alias{aoiextent}
\title{Find the Extent of an Area of Interest}
\usage{
aoiextent(sfid, sfdir, buff = 20)
}
\arguments{
\item{sfid}{A string corresponding to a shapefile filename; can be a shorthand plot ID, e.g. 'SG-NES2'}

\item{buff}{An integer indicating the buffer width around the AOI to include in the clip; default is 20 m}
}
\value{
A list describing the extent of an area of interest in the form \verb{[xmin, xmax, ymin, ymax]}
#' @examples
wf <- aoiextent(wfarrays, 'SG-NES2')
for w in wf:
print(w\link{1:10})
}
\description{
This function allows you to clip waveform data ingested from the ENVI binary format to the extent of a local area of interest.
}
\keyword{geometry}
\keyword{waveform,}
