% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.gauss.R
\name{build.gauss}
\alias{build.gauss}
\title{Build Guassian}
\usage{
build.gauss(x, tbins = 500)
}
\arguments{
\item{x}{a dataframe of the Gaussian parameters}

\item{tbins}{an integer indicating the number of time bins (or x-values) to build along}
}
\description{
The function allows you to build a sum of Gaussian functions from peak-wise parameters
}
