% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipwf.R
\name{clipfun_}
\alias{clipfun_}
\title{Clip Binary Waveform Data}
\usage{
clipfun_(wf, geol, aoi, buff = 20)
}
\arguments{
\item{wf}{An array of waveform data, from one of: return pulse, outgoing pulse, system impulse response, or outgoing impulse response}

\item{geol}{An array of waveform geolocation data}

\item{aoi}{A list describing the extent of an area of interest in the form \verb{[xmin, xmax, ymin, ymax]}}

\item{in_arrays}{A list of input arrays, including the waveform and geolocation}
}
\value{
A list with three components, \code{outgoing}, \code{return}, and \code{geolocation}, each corresponding to a clipped data array
#' @examples
wf <- clipwf(wfarrays, 'SG-NES2')
for w in wf:
print(w\link{1:10})
}
\description{
This function allows you to clip waveform data ingested from the ENVI binary format to the extent of a local area of interest.
}
\keyword{waveform}
