% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv.apply.R
\name{deconv.apply}
\alias{deconv.apply}
\title{Apply Waveform Deconvolution to N Waveforms}
\usage{
deconv.apply(
  rawarray,
  subarray,
  method = "Gold",
  np = 2,
  rescale = T,
  small_paras = list(c(30, 2, 1.8, 30, 1.2, 2)),
  large_paras = list(c(40, 4, 1.8, 40, 4, 1.8))
)
}
\arguments{
\item{rawarray}{is a list of the full set of waveform output from the ingest function. Includes return, outgoing, system impulse response and outgoing impulse response arrays, if all available.}

\item{subarray}{is either the full set of waveforms or a subset, the output of the clipwf function}

\item{method}{is the deconvolution function using either Gold or Richardson-Lucy algorithm. method=c("Gold","RL"). Default is method=c("Gold").}

\item{np}{is a threshold value which specifies when small_paras (fewer iterations and repetitions) or large_paras (more iterations and repetitions) should be applied in the deconvolution, given the number of estimated peaks in the waveform. More complex waveforms require more iterations and more repetitions to converge. Default is 2 peaks.}

\item{rescale}{determines whether to rescale the waveform intensity. Returns are rescaled to the minimum waveform intensity to conduct rescaling. Default is TRUE.}
}
\value{
The deconvolved return waveform
}
\description{
This function applies a choice of Gold or Richardson-Lucy deconvolution algorithm to a set of waveforms of dim(M x N).
}
\examples{
decon <- deconv.apply(wfarrays, cliparrays, method = 'Gold', np=2, rescale = F)

# Check dims of result
dim(decon.dt)

# Count number of weird returns
length(decon.dt[rowSums(decon.dt!=0)])

# Plot some deconvolved returns to check
plot(seq(500), decon[1:500,16], type = 'l')
lines(seq(500), return1[16][[1]], type = 'l')
}
\keyword{Gold,}
\keyword{Richardson-Lucy}
\keyword{deconvolution,}
\keyword{waveform,}
