% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv.clean.R
\name{deconv.clean}
\alias{deconv.clean}
\title{QA and Clean Results of Deconvolution}
\usage{
deconv.clean(de)
}
\arguments{
\item{de}{is a data.table of deconvolved waveform results with popential errors}
}
\value{
The cleaned, deconvolved return waveform
}
\description{
This function finds NaN returns in a set of deconvolved waveforms and infills with 0 or mean of surrounding values
}
\keyword{Gold,}
\keyword{Richardson-Lucy}
\keyword{deconvolution,}
\keyword{waveform,}
