% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decom.clean.R
\name{gfitesc}
\alias{gfitesc}
\title{Clean Error-Generating Waveforms for Decomposition}
\usage{
gfitesc(x)
}
\arguments{
\item{x}{is a list of waveform arrays, such as that resulting from use of the \code{rwaveform::ingest} function}
}
\value{
a list of cleaned/trimmed waveform arrays
}
\description{
This function finds returns in a set of raw or deconvolved waveforms that are likely to generate errors in decomposition and removes them from the set.
}
\examples{

}
\keyword{Gaussian}
\keyword{LiDAR,}
\keyword{decomposition}
\keyword{waveform,}
