% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_parallel.R
\name{message_parallel_}
\alias{message_parallel_}
\title{Apply Functions in Parallel with Safe Exit and Error Handling}
\usage{
message_parallel_(...)
}
\arguments{
\item{X}{is a list of iterables}

\item{FUN}{is an object pointing to a function to be applied in parallel}

\item{clip}{specifies whether to clip the waveform to a particular geometry before processing. Default is not to clip (False)}

\item{aoi}{is list describing the extent of an area of interest in the form \verb{[xmin, xmax, ymin, ymax]}}
}
\value{
a data.frame of points with return parameters and geolocation information
}
\description{
This function allows for generic parallel processing using other functions in the package
}
\examples{

}
\keyword{parallelization}
