% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npeaks.apply.R
\name{npeaks.apply}
\alias{npeaks.apply}
\title{Apply Peak Counter to N Waveforms}
\usage{
npeaks.apply(wfarray, drop = c(0, 0), smooth = F, thres = 0.2)
}
\arguments{
\item{wfarray}{is the array of waveform intensities.}

\item{drop}{is a tuple indicating first and last elements of a range of values we should ignore (e.g., index or non-intensity info). Default is c(0,0), which means use the all input data.}

\item{smooth}{is tell whether you want to smooth the waveform to reduce the effect of some obvious noise. Default is TRUE.}

\item{threshold}{is to determine if the detected peak is the real peak whose intensity should be higher than threshold*maximum intensity. Default is 0.2.}
}
\value{
A list of the number of peaks in each waveform vector
}
\description{
This function allows you to quickly count the number of waveform peaks in an array of N waveforms
}
\examples{
np <- deconv.apply(wfarrays, cliparrays, method = 'Gold', np=2, rescale = F)

}
\keyword{deconvolution,}
\keyword{peaks}
\keyword{waveform,}
