% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakfix.R
\name{peakfix}
\alias{peakfix}
\title{Fix waveform peaks}
\usage{
peakfix(y)
}
\arguments{
\item{y}{is a vector of return intensity values}
}
\value{
the peak-fixed vector
}
\description{
This function finds problematic peaks in a waveform and introduces a minor correction to improve Gaussian curve fitting. After deconvolution, some first peaks in the waveform jump from 0 to the maximum value with no ramp-up. A Gaussian fit can't accommodate this, so when it happens we introduce a value in the prior time bin at 0.99x the first non-zero return value. This creates a shape that a Gaussian can be fit to without invalidating the data.
}
\keyword{deconvolution,}
\keyword{peaks}
\keyword{waveform,}
