% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictloss.R
\name{predictloss}
\alias{predictloss}
\title{Individual Tree Detection}
\usage{
predictloss(inventorydir, shapedir, predictrees, aoi)
}
\arguments{
\item{predictrees}{is a a dataframe of predicted trees with ID, X, Y, and Z values}

\item{aoi}{is a string indicating a plot ID for subset processing}
}
\value{
a value describing total loss in matching field to predicted trees over an area of interest
}
\description{
This function applies a given tree detection algorithm to find individual trees in a LAS/non-LAS point cloud
}
\examples{
sgtrees <- itcdelineate(testlas, aoi)

}
\keyword{RMSE}
\keyword{cloud,}
\keyword{detection,}
\keyword{individual}
\keyword{lidar,}
\keyword{loss,}
\keyword{point}
\keyword{tree}
