% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processwf2.R
\name{process_wf}
\alias{process_wf}
\title{Deconvolve, Decompose, and Geolocate Waveforms}
\usage{
process_wf(fp, logpath, outdir)
}
\arguments{
\item{fp}{is string pointing to the filepath of a directory containing one waveform binary package. The package should typically contain, at minimum, header (.hdr) and binary data files describing:
\itemize{
\item observation parameters
\item geolocation
\item outgoing pulse
\item return pulse
\item impulse response
\item outgoing impulse response
}}

\item{clip}{specifies whether to clip the waveform to a particular geometry before processing. Default is not to clip (False)}

\item{aoi}{is list describing the extent of an area of interest in the form \verb{[xmin, xmax, ymin, ymax]}}
}
\value{
a data.frame of points with return parameters and geolocation information
}
\description{
This function combines other functions in the rwaveform package to create a full processing workflow
}
\examples{

}
\keyword{LiDAR}
\keyword{waveform,}
