% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pts2las.R
\name{pts2las}
\alias{pts2las}
\title{Convert Return Intensity Points to LAS}
\usage{
pts2las(infile, outpath = NULL)
}
\arguments{
\item{aoi}{is a string indicating a plot ID for subset processing}
}
\value{
a LAS point cloud dataset
}
\description{
This function applies an iterative Gaussian decomposition to a set of waveforms of dim (MxN).
}
\examples{
testlas <- pts2las(aoi)

}
\keyword{cloud,}
\keyword{las}
\keyword{lidar,}
\keyword{point}
