% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{readbinary_}
\alias{readbinary_}
\title{Ingest Binary Waveform Data}
\usage{
readbinary_(datafile, headerfile)
}
\arguments{
\item{flightpath}{A string pointing to the filepath of a directory containing one waveform binary package. The package should typically contain, at minimum, header (.hdr) and binary data files describing:
\itemize{
\item observation parameters
\item geolocation
\item outgoing pulse
\item return pulse
\item impulse response
\item outgoing impulse response
}}
}
\value{
A list with five components, \code{out}, \code{re}, \code{geol}, \code{sysir}, and \code{outir}, each representing the arrays contained in the binary package
#' @examples
wf <- ingest('./Data/2018_CRBU_1_2018061214_FL002-001')
for w in wf:
print(w\link{1:10})
}
\description{
This function allows you to ingest waveform data in the ENVI binary format provided by NEON Airborne Observing Platform acquisitions.
}
\keyword{waveform}
