% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itd.R
\name{saferead_}
\alias{saferead_}
\title{Individual Tree Detection}
\usage{
saferead_(x)
}
\arguments{
\item{nlas}{is a ground-normalized las dataset}

\item{aoi}{is a string indicating a plot ID for subset processing}
}
\value{
a dataframe of predicted trees with ID, X, Y, and Z values
}
\description{
This function applies a given tree detection algorithm to find individual trees in a LAS/non-LAS point cloud
}
\examples{
sgtrees <- itcdelineate(testlas, aoi)

}
\keyword{cloud,}
\keyword{detection}
\keyword{individual}
\keyword{las,}
\keyword{lidar,}
\keyword{point}
\keyword{tree}
